requirements.txt
PyYAML==6.0
rich==12.0.0
